Move pedido deletion endpoint to PedidoViewController
