+++ src/main/java/com/proyecto/pis/proyecto_pis/Controller/PedidoViewController.java
@@ -2,10 +2,10 @@ package com.proyecto.pis.proyecto_pis.Controller;
 
 import com.proyecto.pis.proyecto_pis.repository.PedidoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 @RequestMapping("/admin")
@@ -24,5 +24,22 @@ public class PedidoViewController {
     public String obtenerListaPedidos(Model model) {
         model.addAttribute("pedidos", pedidoRepository.findAll());
         return "html/pedidos_admin :: pedidosLista";
+    }
+    
+    @DeleteMapping("/pedido/eliminar/{id}")
+    @ResponseBody
+    public ResponseEntity<?> eliminarPedido(@PathVariable Long id) {
+        try {
+            if (!pedidoRepository.existsById(id)) {
+                return ResponseEntity.badRequest().body("{\"status\":\"error\",\"message\":\"Pedido no encontrado\"}");
+            }
+            
+            pedidoRepository.deleteById(id);
+            System.out.println("Pedido eliminado exitosamente: ID " + id);
+            return ResponseEntity.ok().body("{\"status\":\"success\",\"message\":\"Pedido completado y eliminado exitosamente\"}");
+        } catch (Exception e) {
+            System.out.println("Error al eliminar pedido: " + e.getMessage());
+            return ResponseEntity.badRequest().body("{\"status\":\"error\",\"message\":\"Error al eliminar pedido: " + e.getMessage() + "\"}");
+        }
     }
 }
